Add additional Docker support
